* doc/autoconf.texi (Special Shell Variables): Mention ENV, MAIL, MAILPATH, PS1, PS2, PS4.  Index PWD.
* lib/m4sugar/m4sh.m4 (AS_SHELL_SANITIZE): Unset ENV, MAIL,
MAILPATH and set PS1, PS2, PS4 to default values, to work
around bug in pre-3.0 UWIN ksh reported by Bruce Lilly.
For LC_ALL etc, first try to set to "C" as POSIX requires and as
the Autoconf documentation specifies; fall back to "unset" only if
this fails.  Use a shell for-loop for this rather than an m4 loop,
to shorten the output script.